Created addContactFragment and custom action bar
Created a new fragment called AddContactFragment along with the XML file for it. XML file have 2 text views for "Add Contact" and "Delete Chats". Created a new custom action bar in the activity_main.xml and modified the MainActivity.kt so that it hides the default action bar. Added a new drawable for the back arrow icon and added color entry for text views background color in AddContactFragment.