Implemented a default view for when the recycler view is empty
When the MainFragment is empty, it will display the empty_contact.xml to the user. Updated fragment_main in order to add the empty_contact.xml view. Created NoContactObserver in order to observe the recycler view so that it will display the custom view if empty. Assigned NoContactObserver to the recycler view adapter in MainFragment.